+++ example/lib/main.dart
@@ -4,9 +4,13 @@ import 'package:flutter_flavor/flutter_flavor.dart';
 import 'my_app.dart';
 
 void main() {
-  FlavorConfig(location: BannerLocation.topEnd, variables: {
-    "counter": 0,
-    "baseUrl": "https://www.example1.com",
-  });
+  FlavorConfig(
+    color: Colors.red,
+    location: BannerLocation.topStart,
+    variables: {
+      "counter": 0,
+      "baseUrl": "https://www.example1.com",
+    },
+  );
   return runApp(MyApp());
 }