+++ example/lib/my_app.dart
@@ -5,12 +5,16 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return FlavorBanner(
+      color: Colors.blue,
+      location: BannerLocation.topEnd,
       child: MaterialApp(
         title: 'Flutter Demo',
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        home: MyHomePage(title: 'Flutter Demo Home Page'),
+        home: MyHomePage(
+          title: 'Flutter Demo Home Page',
+        ),
       ),
     );
   }